seperated member js and addingmember component
